+++ gwcelery/tests/test_tasks_first2years.py
@@ -1,5 +1,5 @@
 import io
-from unittest.mock import patch
+from unittest.mock import call, patch
 
 from glue.ligolw import utils
 from glue.ligolw import lsctables
@@ -39,8 +39,13 @@ def test_upload_event(mock_upload, mock_create_event):
     coinc = pick_coinc()
     psd = pkg_resources.resource_string(
         __name__, '../data/first2years/2016/psd.xml.gz')
+    ranking_data = pkg_resources.resource_string(
+        __name__, 'data/ranking_data_G322589.xml.gz')
 
     upload_event()
+
     mock_create_event.assert_called_once_with(coinc, 'MDC', 'gstlal', 'CBC')
-    mock_upload.assert_called_once_with(
-        psd, 'psd.xml.gz', 'M1234', 'Noise PSD', ['psd'])
+    mock_upload.assert_has_calls([
+        call(psd, 'psd.xml.gz', 'M1234', 'Noise PSD', ['psd']),
+        call(ranking_data, 'ranking_data.xml.gz', 'M1234', 'Ranking data')
+    ])